+++ FluffyEngine/Texture2D.h
@@ -13,7 +13,7 @@ public:
 	explicit Texture2D(SDL_Texture* texture);
 	~Texture2D();
 
-	glm::ivec2 GetSize() const;
+	glm::vec2 GetSize() const;
 
 	Texture2D(const Texture2D&) = delete;
 	Texture2D(Texture2D&&) = delete;
